Add tests on REST blueprints
