+++ packages/acceptance-tests/src/docs/api-section/rest-blueprints/product.entity.ts
@@ -1,0 +1,12 @@
+import { BaseEntity, Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
+
+@Entity()
+export class Product extends BaseEntity {
+
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column()
+  text: string;
+
+}
